feat: enable processing of DataView Inline tasks
This change adds a new feature to allow DataView Inline fields to be processed and added to the tasks in the obsidian_markdown_tasks table. They are exposed as additional columns so you can use them in queries. A new setting has been added to enable this feature. It is disabled by default.

This resolves issue Support arbitrary task inline fields #11