fix: upsert to task and list on note change and add content back to note
This change fixes the missing content value from the notes table, in addition the changed event was not inserting new tasks or list items if added to the page. The event was updated to make this a upsert event for both tasks and lists to correct this. The Documentation is also updated for the new tables.